src/interface/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grimoire-py"
version = "0.1.0"
description = "Grimoire Python Package - Python interface with CLI, MCP server, and REST API for Grimoire knowledge-based OS"
authors = [
    {name = "Grimoire", email = "grimoire@example.com"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "janus-swi @ git+https://github.com/SWI-Prolog/packages-swipy.git",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.scripts]
grimoire-cli = "grimoire.cli:main"
grimoire-mcp = "grimoire.mcp:main"
grimoire-http = "grimoire.http:main"

[tool.setuptools]
packages = ["grimoire"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]